feat: add transaction to handle store and update data

+++ src/module/user/models/update.model.ts
@@ -1,20 +1,25 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ResponseMessage } from 'src/common/message/message.enum';
-import { Repository, UpdateResult } from 'typeorm';
+import { DataSource, EntityManager, Repository, UpdateResult } from 'typeorm';
 import { UserEntity } from '../entities/user.entity';
 import { UserAllergyEntity } from '../entities/user-allergy.entity';
 
 @Injectable()
 export class UpdateModel {
 	constructor(
+		private readonly dataSource: DataSource,
 		@InjectRepository(UserEntity)
 		private readonly UserRepository: Repository<UserEntity>,
 		@InjectRepository(UserAllergyEntity)
 		private readonly UserAllergyRepository: Repository<UserAllergyEntity>,
 	) {}
 
-	async updateUser({ params, uid }): Promise<UpdateResult> {
+	async updateUser(
+		params,
+		uid: string,
+		em: EntityManager = this.dataSource.manager,
+	): Promise<UpdateResult> {
 		try {
 			const user = await this.UserRepository.createQueryBuilder('user')
 				.where('uid = :uid', { uid: uid })
@@ -38,14 +43,12 @@ export class UpdateModel {
 		}
 	}
 
-	async deleteExistingAllergy(uid: string) {
+	async deleteExistingUserAllergy(
+		uid: string,
+		em: EntityManager = this.dataSource.manager,
+	) {
 		try {
-			const query = this.UserAllergyRepository.createQueryBuilder('ua')
-				.delete()
-				.from('user_allergy')
-				.where('user_id = :user_id', { user_id: uid });
-
-			return await query.execute();
+			return await em.delete(UserAllergyEntity, { user_id: uid });
 		} catch (error) {
 			throw error;
 		}
